refactor[TW19947]: Remove RelationalConstants.DEFAULT_ITEM_ID

+++ plugins/org.eclipse.osee.orcs.core/src/org/eclipse/osee/orcs/core/ds/OrcsData.java
@@ -20,7 +20,7 @@ import org.eclipse.osee.framework.jdk.core.type.Id;
 
 public interface OrcsData<T extends Id> extends HasVersion {
 
-   void setLocalId(Integer localId);
+   void setLocalId(long localId);
 
    Integer getLocalId();
